listitemmanager: Change the way listitems are handled
Instead of immediately removing listitems when they are added/removed or
their widgets become unnecessary, force a manual cleanup. Do this
cleanup in size_allocate().

This way, we can assign space to listitems during allocate and be sure
it survives until the next call to size_allocate(), which allows
tracking multiple changes to the listitems in a single frame, as the
assigned space never collapses.

It also simplifies gridview code, because it allows splitting listitems
so that every listitem represents a rectangular region.

Fixes #2971